[delimitedtext] Move GUI-related code to a separate file
