+++ src/providers/delimitedtext/CMakeLists.txt
@@ -15,6 +15,7 @@ SET (DTEXT_MOC_HDRS
 
 IF (WITH_GUI)
   SET(DTEXT_SRCS ${DTEXT_SRCS}
+    qgsdelimitedtextprovidergui.cpp
     qgsdelimitedtextsourceselect.cpp
   )
   SET(DTEXT_MOC_HDRS ${DTEXT_MOC_HDRS}
